Moved events and aggregate definitions to API
