+++ pkg/api/events/events.proto
@@ -3,7 +3,7 @@ syntax = "proto3";
 package fission.workflows.events;
 option go_package = "events";
 
-import "github.com/fission/fission-workflows/pkg/types/types.proto";
+import "pkg/types/types.proto";
 
 message EventWrapper {
     string any = 1;
